requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
httpx==0.27.2
python-dotenv==1.0.1
openpyxl==3.1.5
apscheduler==3.10.4
python-multipart==0.0.12
